backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
asyncpg
requests
pydantic
pyright
apscheduler
pytest
pytest-cov
httpx
pytest-asyncio
python-jose[cryptography]
passlib[bcrypt]
alembic
PyJWT
